+++ client/src/main/scala/org/http4s/client/middleware/Logger.scala
@@ -3,6 +3,7 @@ package client
 package middleware
 
 import cats.effect._
+import cats.implicits._
 import fs2._
 import org.http4s.util.CaseInsensitiveString
 
@@ -52,25 +53,22 @@ object Logger {
       message.bodyAsText(charset.getOrElse(Charset.`UTF-8`))
     } else if (logBody) {
       message.body
-        .fold(new StringBuilder)((sb, b) => sb.append(java.lang.Integer.toHexString(b & 0xff)))
-        .map(_.toString)
+        .map(b => java.lang.Integer.toHexString(b & 0xff))
     } else {
       Stream.empty.covary[F]
     }
 
     val bodyText = if (logBody) {
-      bodyStream.fold("")(_ + _).map(text => s"""body="$text"""")
+      bodyStream.compile.string
+        .map(text => s"""body="$text"""")
     } else {
-      Stream("").covary[F]
+      F.pure("")
     }
 
-    if (!logBody && !logHeaders) F.unit
-    else {
-      bodyText
-        .map(body => s"$prelude $headers $body")
-        .evalMap(text => log(text))
-        .compile
-        .drain
-    }
+    def spaced(x: String): String = if (x.isEmpty) x else s" $x"
+
+    bodyText
+      .map(body => s"$prelude${spaced(headers)}${spaced(body)}")
+      .flatMap(log)
   }
 }